src/enhancements/quantum_inspired_optimization/quantum_inspired_optimization.py
import numpy as np
from qiskit import QuantumCircuit
from qiskit_aer import AerSimulator
from scipy.optimize import minimize
from qiskit.circuit import Parameter
from typing import List, Tuple, Union, Dict
import logging
import re

class QuantumInspiredOptimizer:
    def __init__(self, num_qubits: int, num_shots: int = 1000, p: int = 1):
        self.num_qubits = num_qubits
        self.num_shots = num_shots
        self.p = p
        self.backend = AerSimulator()

    def convert_state_to_binary(self, state):
        if isinstance(state, int):
            return format(state, f'0{self.num_qubits}b')
        elif isinstance(state, str):
            cleaned_state = ''.join(filter(lambda x: x in '01', state))
            return format(int(cleaned_state, 2), f'0{self.num_qubits}b')
        elif isinstance(state, (list, tuple)):
            return ''.join(str(bit) for bit in state)
        else:
            raise ValueError(f"Unsupported type for state: {type(state)}")

    def validate_binary_string(self, binary_str):
        if not all(char in '01' for char in binary_str):
            raise ValueError(f"Invalid binary string: {binary_str}")
        return binary_str

    def create_circuit(self, beta: List[float], gamma: List[float]) -> QuantumCircuit:
        """
        Create a parameterized quantum circuit for QAOA.
        """
        if len(beta) != self.p or len(gamma) != self.p:
            raise ValueError(f"Expected {self.p} parameters for beta and gamma, got {len(beta)} and {len(gamma)} respectively.")

        circuit = QuantumCircuit(self.num_qubits, self.num_qubits)

        # Initial state
        circuit.h(range(self.num_qubits))

        # QAOA layers
        for p in range(self.p):
            self._apply_cost_hamiltonian(circuit, gamma[p])
            self._apply_mixing_hamiltonian(circuit, beta[p])

        # Measurement
        circuit.measure_all()

        return circuit

    def _apply_cost_hamiltonian(self, circuit: QuantumCircuit, gamma: float):
        """
        Apply the cost Hamiltonian e^(-iγ H_C).
        """
        # Implement a simple cost Hamiltonian
        for i in range(self.num_qubits):
            circuit.rz(2 * gamma, i)

    def _apply_mixing_hamiltonian(self, circuit: QuantumCircuit, beta: float):
        """
        Apply the mixing Hamiltonian e^(-iβ H_B).
        """
        for i in range(self.num_qubits):
            circuit.rx(2 * beta, i)

    def optimize_memory_task(self, task_complexity: float) -> Dict[str, Union[int, float, str]]:
        """
        Optimize the memory task using QAOA.
        """
        def objective_function(params):
            beta, gamma = params[:self.p], params[self.p:]
            circuit = self.create_circuit(beta, gamma)
            try:
                counts = self.backend.run(circuit, shots=self.num_shots).result().get_counts()
                return -self._compute_expectation(counts)  # We negate because we want to maximize
            except Exception as e:
                logging.error(f"Error in objective function: {str(e)}")
                return float('inf')  # Return a large value to indicate failure

        initial_params = np.random.rand(2 * self.p)
        try:
            result = minimize(objective_function, initial_params, method='COBYLA')
        except Exception as e:
            logging.error(f"Optimization failed: {str(e)}")
            return {'error': 'Optimization failed'}

        optimized_params = result.x
        optimized_circuit = self.create_circuit(optimized_params[:self.p], optimized_params[self.p:])
        try:
            counts = self.backend.run(optimized_circuit, shots=self.num_shots).result().get_counts()
        except Exception as e:
            logging.error(f"Error running optimized circuit: {str(e)}")
            return {'error': 'Failed to run optimized circuit'}

        optimized_state = max(counts, key=counts.get)
        strategy_dict = self.map_state_to_strategy(optimized_state, task_complexity)

        return {
            'state': strategy_dict['state'],
            'task_complexity': strategy_dict['task_complexity'],
            'selected_strategy': strategy_dict['selected_strategy']
        }

    def _compute_expectation(self, counts: dict) -> float:
        """
        Compute the expectation value of the cost Hamiltonian.
        """
        expectation = 0.0
        total_counts = sum(counts.values())
        for state, count in counts.items():
            # Remove spaces and ensure only binary digits are processed
            binary_state = ''.join(bit for bit in state if bit in ('0', '1'))
            # Implement a simple cost function
            cost = sum(int(bit) for bit in binary_state)
            expectation += cost * count / total_counts
        return expectation

    def map_state_to_strategy(self, state: Union[int, str, List[int], Tuple[int]], task_complexity: float) -> Dict[str, Union[int, float, str]]:
        """
        Maps the optimized quantum state to a concrete memory enhancement strategy.
        """
        logging.info(f"Mapping state to strategy with state: {state}, task_complexity: {task_complexity}")

        if not isinstance(state, (int, str, list, tuple)):
            raise ValueError(f"Invalid state type: {type(state)}. Expected int, str, list, or tuple.")

        try:
            state_str = self.convert_state_to_binary(state)
            state_str = self.validate_binary_string(state_str)
        except (TypeError, ValueError) as e:
            logging.exception("Error during state conversion or validation.")
            raise ValueError(f"Invalid state: {state}. Error: {str(e)}")

        logging.debug(f"Validated binary string: '{state_str}'")

        try:
            state_int = int(state_str, 2)
            logging.debug(f"Converted to integer: {state_int}")
        except ValueError as ve:
            logging.exception("Error converting binary string to integer.")
            raise ValueError(f"Invalid binary string: {state_str}. Error: {str(ve)}")

        strategies = [
            "Chunking", "Mnemonic Devices", "Mind Mapping", "Spaced Repetition",
            "Method of Loci", "Elaborative Rehearsal", "Dual Coding", "Active Recall"
        ]

        strategy_index = (state_int + int(task_complexity * 10)) % len(strategies)
        selected_strategy = strategies[strategy_index]

        result = {
            'state': state_int,
            'task_complexity': task_complexity,
            'selected_strategy': selected_strategy
        }

        logging.info(f"Mapped strategy: {result}")
        return result

def quantum_inspired_memory_optimization(task_description: str, task_complexity: float) -> str:
    """
    Optimize memory enhancement strategy using quantum-inspired optimization.

    Args:
    task_description (str): Description of the memory task.
    task_complexity (float): Complexity of the task, ranging from 0 to 1.

    Returns:
    str: Optimized memory enhancement strategy.
    """
    num_qubits = 3  # Adjust based on the complexity of the optimization problem
    optimizer = QuantumInspiredOptimizer(num_qubits, p=2)  # Using 2 QAOA layers
    optimized_strategy = optimizer.optimize_memory_task(task_complexity)

    return f"For the task '{task_description}' with complexity {task_complexity}, " \
           f"the optimized memory enhancement strategy is: {optimized_strategy}"

# Example usage
if __name__ == "__main__":
    task = "Memorize the periodic table"
    complexity = 0.8
    result = quantum_inspired_memory_optimization(task, complexity)
    print(result)
